+++ plugins/modules/repositories.py
@@ -14,7 +14,7 @@ author: "Artem Goncharov (@gtema)"
 description:
   - Manages repository options
 options:
-  token:
+  root:
     description: GitHub token
     type: str
     required: True
@@ -26,12 +26,12 @@ RETURN = '''
 EXAMPLES = '''
 '''
 
-from ansible_collections.opentelekomcloud.gitcontrol.plugins.module_utils.git import GitBase
+from ansible_collections.opentelekomcloud.gitcontrol.plugins.module_utils.github import GitHubBase
 
 
-class Repo(GitBase):
+class Repo(GitHubBase):
     argument_spec = dict(
-        token=dict(type='str', required=True, no_log=True)
+        root=dict(type='str', required=False),
     )
     module_kwargs = dict(
         supports_check_mode=True
