+++ tox.ini
@@ -46,7 +46,7 @@ passenv = *
 deps =
     {[testenv]deps}
 commands =
-    ansible-lint 
+    ansible-lint -vvv
 
 [testenv:linters-2.9]
 passenv = {[testenv:linters]passenv}
